requirements.txt
pandas
requests
apache-airflow
psycopg2-binary
duckdb
pyarrow
python-dotenv
